+++ TeamProject/api/user.py
@@ -7,102 +7,11 @@ from models import User, db
 from werkzeug.security import generate_password_hash, check_password_hash
 from datetime import datetime
 from werkzeug.utils import secure_filename
+from controllers.user_controller import *
 import re
 
 # 임시
 from flask import g
-
-# 토큰을 받고 해당 토큰이 관리잔지 유저인지 확인하는 api
-@api.route("/who_are_you")
-@jwt_required
-def who_are_you():
-    data = request.get_json()
-    id = data.get("id")
-    check_user_token = get_jwt_identity()
-    if check_user_token == "GM":
-        return jsonify(True)
-    else:
-        check_user = User.query.filter(User.id == id).first()
-        if check_user.userid == check_user_token:
-            return jsonify(True)
-        else:
-            return jsonify(False)
-
-
-# 이미지 기본 설정
-ALLOWED_EXTENSIONS = {"png", "jpg", "jpeg", "gif"}
-UPLOAD_FOLDER = "static/img/profile_img"
-
-
-def allowed_file(file):
-    check = 1
-    if (
-        file.filename.rsplit(".", 1)[1].lower() not in ALLOWED_EXTENSIONS
-        or "." not in file.filename
-    ):
-        check = 0
-
-    return check
-
-
-# 프로필 사진 업로드
-@api.route("/profile_img_upload/<id>", methods=["POST"])
-@jwt_required
-def projile_img_upload(id):
-    profile_img = request.files["profile_img"]  # 프로필 사진 받아도 되고 안받아도 됨
-    # POST request에 파일 정보가 있는지 확인
-    print(request.files)
-
-    if "profile_img" not in request.files:
-        print("No file part")
-        return redirect("api/progile_img_upload/<id>")
-
-    # 만약 유저가 파일을 고르지 않았을 경우
-    if profile_img.filename == "":
-        print("No selected file")
-        return redirect("api/progile_img_upload/<id>")
-
-    # 프로필 사진 이름 유저 테이블에 삽입 및 저장
-    if profile_img and allowed_file(profile_img):  # 프로필 이미지 확장자 확인
-        suffix = datetime.now().strftime("%y%m%d_%H%M%S")
-        filename = "_".join(
-            [profile_img.filename.rsplit(".", 1)[0], suffix]
-        )  # 중복된 이름의 사진을 받기위해서 파일명에 시간을 붙임
-        extension = profile_img.filename.rsplit(".", 1)[1]
-        filename = secure_filename(f"{filename}.{extension}")
-
-        user = User.query.filter(User.id == id).first()
-        user.profile_img = filename
-        db.session.add(user)
-        db.session.commit()
-
-        profile_img.save(os.path.join(UPLOAD_FOLDER, filename))
-        return {"msg": "프로필 사진 등록 완료"}, 201
-    return {"msg": "프로필 사진 등록이 안댓다"}, 404
-
-
-def pwd_check(password):
-    result = {}
-    if len(password) < 6 or len(password) > 12:
-        result = {"error": "비밀번호는 6자리 이상 12자리 이하입니다.", "error_code": 403}  # 데이터 유효성 검사
-        return result
-    if len(re.findall("[^a-zA-Z0-9]", password)) == 0:
-        result = {"error": "비밀번호에 특수문자가 포함되어 있어야 합니다.", "error_code": 403}  # 데이터 유효성 검사
-        return result
-    return result
-
-
-def email_check(email):
-    result = {}
-    reg = re.findall("^[a-z0-9]{2,}@[a-z]{2,}\.[a-z]{2,}$", email)
-    if len(reg) == 0:
-        result = {"error": "이메일 형식이 옳지 않습니다.", "error_code": 403}  # 데이터 유효성 검사
-        return result
-    if User.query.filter(User.email == email).first():
-        result = {"error": "이미 가입이 된적 있는 이메일입니다.", "error_code": 409}  # 중복 오류 코드
-        return result
-    return result
-
 
 @api.route("/sign_up", methods=["POST"])  # 회원 가입 api 및 임시로 데이터 확인api
 def sign_up():
@@ -230,12 +139,6 @@ def user_info():
         }
         return jsonify(access_user_info), 201  # 모든 사용자정보 반환
 
-    # ------------------------------신경 안써도댐------------------------------
-    #  res_users = {}
-    #  for user in users:# 반복문을 돌면서 직렬화된 변수를 넣어서 새로운 리스트를 만든다.
-    #      res_users.append(user.serialize)
-    #  return jsonify(res_users)
-    # ------------------------------------------------------------------------
 
 
 # 아이디 삭제, 수정, id(primary key)값에 따른 정보확인
