+++ TeamProject/config.py
@@ -9,3 +9,5 @@ SECRET_KEY = "123456789"
 JWT_SECRET_KEY = "1232132152142"
 ADMIN_ID = "GM"
 ADMIN_PW = "1234"
+ALLOWED_EXTENSIONS = {"png", "jpg", "jpeg", "gif"}
+UPLOAD_FOLDER = "static/img/profile_img"
